+++ src/main/kotlin/io/hhplus/concertreservationservice/common/filter/LoggingFilter.kt
@@ -1,4 +1,4 @@
-package io.hhplus.concertreservationservice.infrastructure.filter
+package io.hhplus.concertreservationservice.common.filter
 
 import jakarta.servlet.FilterChain
 import jakarta.servlet.ServletException
