+++ src/main/kotlin/io/hhplus/concertreservationservice/common/filter/RequestWrapper.kt
@@ -1,4 +1,4 @@
-package io.hhplus.concertreservationservice.infrastructure.filter
+package io.hhplus.concertreservationservice.common.filter
 
 import jakarta.servlet.ReadListener
 import jakarta.servlet.ServletInputStream
