+++ src/main/kotlin/io/hhplus/concertreservationservice/common/interceptor/RequestResponseLoggingInterceptor.kt
@@ -1,4 +1,4 @@
-package io.hhplus.concertreservationservice.infrastructure.interceptor
+package io.hhplus.concertreservationservice.common.interceptor
 
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
